backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.1
boto3==1.20.26
botocore==1.23.54
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.7
django-anymail==10.2
django-cors-headers==3.14.0
django-environ==0.12.0
django-extensions==3.2.3
django-import-export==4.0.5
django-jazzmin==2.6.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==10.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==0.10.0
kombu==5.4.2
marshmallow==3.20.1
mysqlclient==2.2.7
packaging==23.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.5.2
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
